Struct libgweather::Info

#[repr(transparent)]
pub struct Info { /* private fields */ }

GWeatherInfo provides a handy way to access weather conditions and forecasts from a Location, aggregating multiple different web services.

It includes also astronomical data such as sunrise times and moon phases.

Implementations

impl Info

pub fn new(location: Option<&Location>) -> Info

Builds a new GWeatherInfo that will provide weather information about the given location.

In order to retrieve the weather information, you will need to enable the desired providers and then call update(). If you want to be notified of the completion of the weather information update, you should connect to the signal::Info::updated signal before updating the GWeatherInfo instance.

location

the desidered location

Returns

a new weather information instance Builds a new GWeatherInfo that will provide weather information about the given location.

In order to retrieve the weather information, you will need to enable the desired providers and then call update(). If you want to be notified of the completion of the weather information update, you should connect to the signal::Info::updated signal before updating the GWeatherInfo instance.

location

the desidered location

Returns

a new weather information instance

pub fn builder() -> InfoBuilder

Creates a new builder-pattern struct instance to construct Info objects.

This method returns an instance of InfoBuilder which can be used to create Info objects.

pub fn abort(&self)

pub fn apparent(&self) -> GString

pub fn application_id(&self) -> GString

Get the application ID of the application fetching the weather.

Returns

the application ID Get the application ID of the application fetching the weather.

Returns

the application ID

pub fn attribution(&self) -> GString

Some weather services require the application showing the data to include an attribution text, possibly including links to the service website. This must be shown prominently toghether with the data.

Returns

the required attribution text, in Pango markup form, or None if not required Some weather services require the application showing the data to include an attribution text, possibly including links to the service website. This must be shown prominently toghether with the data.

Returns

the required attribution text, in Pango markup form, or None if not required

pub fn conditions(&self) -> GString

pub fn contact_info(&self) -> GString

Get the contact information of the application fetching the weather.

Returns

the contact information Get the contact information of the application fetching the weather.

Returns

the contact information

pub fn dew(&self) -> GString

pub fn enabled_providers(&self) -> Provider

Gets the bitmask of enabled #GWeatherProvider weather providers. Gets the bitmask of enabled #GWeatherProvider weather providers.

pub fn forecast_list(&self) -> Vec<Info>

Returns

list of GWeatherInfo* objects for the forecast. The list is owned by the ‘info’ object thus is alive as long as the ‘info’. The ‘update’ property is the date/time when the forecast info is used for.

Returns

list of GWeatherInfo* objects for the forecast. The list is owned by the ‘info’ object thus is alive as long as the ‘info’. The ‘update’ property is the date/time when the forecast info is used for.

pub fn humidity(&self) -> GString

pub fn icon_name(&self) -> GString

pub fn location(&self) -> Location

pub fn location_name(&self) -> GString

pub fn pressure(&self) -> GString

pub fn sky(&self) -> GString

pub fn sunrise(&self) -> GString

pub fn sunset(&self) -> GString

pub fn symbolic_icon_name(&self) -> GString

pub fn temp(&self) -> GString

pub fn temp_max(&self) -> GString

pub fn temp_min(&self) -> GString

pub fn temp_summary(&self) -> GString

pub fn upcoming_moonphases(&self) -> Option<c_long>

pub fn value_apparent(&self, unit: TemperatureUnit) -> Option<f64>

unit

the desired unit, as a #GWeatherTemperatureUnit

Returns

TRUE is @value is valid, FALSE otherwise.

value

the apparent temperature

unit

the desired unit, as a #GWeatherTemperatureUnit

Returns

TRUE is @value is valid, FALSE otherwise.

value

the apparent temperature

pub fn value_conditions(
&self
) -> Option<(ConditionPhenomenon, ConditionQualifier)>

Fills out @phenomenon and @qualifier with current weather conditions.

Returns

TRUE is out arguments are valid, FALSE otherwise.

phenomenon

a location for a #GWeatherConditionPhenomenon.

qualifier

a location for a #GWeatherConditionQualifier. Fills out @phenomenon and @qualifier with current weather conditions.

Returns

TRUE is out arguments are valid, FALSE otherwise.

phenomenon

a location for a #GWeatherConditionPhenomenon.

qualifier

a location for a #GWeatherConditionQualifier.

pub fn value_dew(&self, unit: TemperatureUnit) -> Option<f64>

unit

the desired unit, as a #GWeatherTemperatureUnit

Returns

TRUE is @value is valid, FALSE otherwise.

value

the dew point

unit

the desired unit, as a #GWeatherTemperatureUnit

Returns

TRUE is @value is valid, FALSE otherwise.

value

the dew point

pub fn value_moonphase(&self) -> Option<(MoonPhase, MoonLatitude)>

Returns

TRUE is @value is valid, FALSE otherwise.

value

the current moon phase

lat

the moon declension

Returns

TRUE is @value is valid, FALSE otherwise.

value

the current moon phase

lat

the moon declension

pub fn value_pressure(&self, unit: PressureUnit) -> Option<f64>

unit

the desired unit, as a #GWeatherPressureUnit

Returns

TRUE if @value is valid, FALSE otherwise.

value

forecasted pressure, expressed in @unit

unit

the desired unit, as a #GWeatherPressureUnit

Returns

TRUE if @value is valid, FALSE otherwise.

value

forecasted pressure, expressed in @unit

pub fn value_sky(&self) -> Option<Sky>

Fills out @sky with current sky conditions.

Returns

TRUE is @sky is valid, FALSE otherwise.

sky

a location for a #GWeatherSky. Fills out @sky with current sky conditions.

Returns

TRUE is @sky is valid, FALSE otherwise.

sky

a location for a #GWeatherSky.

pub fn value_sunrise(&self) -> Option<c_ulong>

Returns

TRUE is @value is valid, FALSE otherwise.

value

the time of sunrise

Returns

TRUE is @value is valid, FALSE otherwise.

value

the time of sunrise

pub fn value_sunset(&self) -> Option<c_ulong>

Returns

TRUE is @value is valid, FALSE otherwise.

value

the time of sunset

Returns

TRUE is @value is valid, FALSE otherwise.

value

the time of sunset

pub fn value_temp(&self, unit: TemperatureUnit) -> Option<f64>

unit

the desired unit, as a #GWeatherTemperatureUnit

Returns

TRUE is @value is valid, FALSE otherwise.

value

the temperature value

unit

the desired unit, as a #GWeatherTemperatureUnit

Returns

TRUE is @value is valid, FALSE otherwise.

value

the temperature value

pub fn value_temp_max(&self, unit: TemperatureUnit) -> Option<f64>

unit

the desired unit, as a #GWeatherTemperatureUnit

Returns

TRUE is @value is valid, FALSE otherwise.

value

the maximum temperature value

unit

the desired unit, as a #GWeatherTemperatureUnit

Returns

TRUE is @value is valid, FALSE otherwise.

value

the maximum temperature value

pub fn value_temp_min(&self, unit: TemperatureUnit) -> Option<f64>

unit

the desired unit, as a #GWeatherTemperatureUnit

Returns

TRUE is @value is valid, FALSE otherwise.

value

the minimum temperature value

unit

the desired unit, as a #GWeatherTemperatureUnit

Returns

TRUE is @value is valid, FALSE otherwise.

value

the minimum temperature value

pub fn value_update(&self) -> Option<c_long>

Note that @value may be 0 if @self has not yet been updated.

Returns

TRUE is @value is valid, FALSE otherwise.

value

the time @self was last updated Note that @value may be 0 if @self has not yet been updated.

Returns

TRUE is @value is valid, FALSE otherwise.

value

the time @self was last updated

pub fn value_visibility(&self, unit: DistanceUnit) -> Option<f64>

unit

the desired unit, as a #GWeatherDistanceUnit

Returns

TRUE if @value is valid, FALSE otherwise.

value

forecasted visibility, expressed in @unit

unit

the desired unit, as a #GWeatherDistanceUnit

Returns

TRUE if @value is valid, FALSE otherwise.

value

forecasted visibility, expressed in @unit

pub fn value_wind(&self, unit: SpeedUnit) -> Option<(f64, WindDirection)>

unit

the desired unit, as a #GWeatherSpeedUnit

Returns

TRUE if @speed and @direction are valid, FALSE otherwise.

speed

forecasted wind speed

direction

forecasted wind direction

unit

the desired unit, as a #GWeatherSpeedUnit

Returns

TRUE if @speed and @direction are valid, FALSE otherwise.

speed

forecasted wind speed

direction

forecasted wind direction

pub fn visibility(&self) -> GString

pub fn weather_summary(&self) -> GString

Returns

a summary for current weather conditions.

Returns

a summary for current weather conditions.

pub fn wind(&self) -> GString

pub fn is_daytime(&self) -> bool

Returns

Whether it is daytime (that is, if the sun is visible) or not at the location and the point of time referred by @self. This is mostly equivalent to comparing the return value of gweather_info_get_value_sunrise() and gweather_info_get_value_sunset(), but it accounts also for midnight sun and polar night, for locations within the Artic and Antartic circles.

Returns

Whether it is daytime (that is, if the sun is visible) or not at the location and the point of time referred by @self. This is mostly equivalent to comparing the return value of gweather_info_get_value_sunrise() and gweather_info_get_value_sunset(), but it accounts also for midnight sun and polar night, for locations within the Artic and Antartic circles.

pub fn is_valid(&self) -> bool

pub fn network_error(&self) -> bool

pub fn next_sun_event(&self) -> i32

pub fn set_application_id(&self, application_id: &str)

Sets the application ID of the application fetching the weather. It is a requirement for using any of the online weather providers.

If the application uses #GApplication, then the application ID will be automatically filled in.

application_id

the application ID to set Sets the application ID of the application fetching the weather. It is a requirement for using any of the online weather providers.

If the application uses #GApplication, then the application ID will be automatically filled in.

application_id

the application ID to set

pub fn set_contact_info(&self, contact_info: &str)

Sets the contact information for the application fetching the weather. It is a requirement for using any of the online weather providers as it allows API providers to contact application developers in case of terms of use breaches.

The contact information should be an email address, or the full URL to an online contact form which weather providers can use to contact the application developer. Avoid using bug tracker URLs which require creating accounts.

contact_info

the contact information for the application Sets the contact information for the application fetching the weather. It is a requirement for using any of the online weather providers as it allows API providers to contact application developers in case of terms of use breaches.

The contact information should be an email address, or the full URL to an online contact form which weather providers can use to contact the application developer. Avoid using bug tracker URLs which require creating accounts.

contact_info

the contact information for the application

pub fn set_enabled_providers(&self, providers: Provider)

Sets the enabled providers for fetching the weather. Note that it is up to the application developer to make sure that the terms of use for each service are respected.

Online providers will not be enabled if the application ID is not set to a valid value.

providers

a bitmask of #GWeatherProvider Sets the enabled providers for fetching the weather. Note that it is up to the application developer to make sure that the terms of use for each service are respected.

Online providers will not be enabled if the application ID is not set to a valid value.

providers

a bitmask of #GWeatherProvider

pub fn set_location(&self, location: Option<&Location>)

Changes the location of the weather report.

Note that this will clear any forecast or current conditions, and you must call update() to obtain the new data.

location

a location for which weather is desired Changes the location of the weather report.

Note that this will clear any forecast or current conditions, and you must call update() to obtain the new data.

location

a location for which weather is desired

pub fn store_cache()

Ensures that any data cached from the network is stored to disk. Calling this is not necessary, as the cache will be saved when the last reference to a #GWeatherInfo will be dropped. On the other hand, it must be called if there is any chance that the application will be closed without unreffing all objects, such as when using a language binding that employs a GC. Ensures that any data cached from the network is stored to disk. Calling this is not necessary, as the cache will be saved when the last reference to a #GWeatherInfo will be dropped. On the other hand, it must be called if there is any chance that the application will be closed without unreffing all objects, such as when using a language binding that employs a GC.

pub fn connect_updated<F: Fn(&Self) + 'static>(&self, f: F) -> SignalHandlerId

This signal is emitted after the initial fetch of the weather data from upstream services, and after every successful call to update(). This signal is emitted after the initial fetch of the weather data from upstream services, and after every successful call to update().

pub fn connect_application_id_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_contact_info_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_enabled_providers_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_location_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

impl Info

pub fn get_update(&self) -> Option<GString>

pub fn update(&self)

Requests a reload of weather conditions and forecast data from enabled network services. This call does no synchronous IO: rather, the result is delivered by emitting the #GWeatherInfo::updated signal. Note that if no network services are enabled, the signal will not be emitted. See #GWeatherInfo:enabled-providers for details. Requests a reload of weather conditions and forecast data from enabled network services. This call does no synchronous IO: rather, the result is delivered by emitting the #GWeatherInfo::updated signal. Note that if no network services are enabled, the signal will not be emitted. See #GWeatherInfo:enabled-providers for details.