Struct libgweather::Location

#[repr(transparent)]
pub struct Location { /* private fields */ }

A GWeatherLocation represents a “location” of some type known to libgweather; anything from a single weather station to the entire world.

See LocationLevel for information about how the hierarchy of locations works.

Implementations

impl Location

pub fn new_detached(
name: &str,
icao: Option<&str>,
latitude: f64,
longitude: f64
) -> Location

Construct a new location from the given data, supplementing any missing information from the static database.

name

the user visible location name

icao

the ICAO code of the location

latitude

the latitude of the location

longitude

the longitude of the location

Returns

the newly created detached location Construct a new location from the given data, supplementing any missing information from the static database.

name

the user visible location name

icao

the ICAO code of the location

latitude

the latitude of the location

longitude

the longitude of the location

Returns

the newly created detached location

pub fn deserialize(&self, serialized: &Variant) -> Location

This call undoes the effect of gweather_location_serialize(), that is, it turns a #GVariant into a #GWeatherLocation. The conversion happens in the context of @self (i.e, for a city or weather station, the resulting location will be attached to a administrative division, country and region as appropriate).

serialized

the #GVariant representing the #GWeatherLocation

Returns

the deserialized location. This call undoes the effect of gweather_location_serialize(), that is, it turns a #GVariant into a #GWeatherLocation. The conversion happens in the context of @self (i.e, for a city or weather station, the resulting location will be attached to a administrative division, country and region as appropriate).

serialized

the #GVariant representing the #GWeatherLocation

Returns

the deserialized location.

pub fn equal(&self, two: &Location) -> bool

pub fn find_by_country_code(&self, country_code: &str) -> Location

Retrieves the country identified by the specified ISO 3166 code, if present in the database.

country_code

a country code

Returns

a country level #GWeatherLocation, or None. Retrieves the country identified by the specified ISO 3166 code, if present in the database.

country_code

a country code

Returns

a country level #GWeatherLocation, or None.

pub fn find_by_station_code(&self, station_code: &str) -> Option<Location>

Retrieves the weather station identifier by @station_code.

Note that multiple instances of the same weather station can exist in the database, and this function will return any of them, so this not usually what you want.

See deserialize() to recover a stored location.

station_code

a 4 letter METAR code

Returns

a weather station level location for the given station code, or NULL if none exists in the database Retrieves the weather station identifier by @station_code.

Note that multiple instances of the same weather station can exist in the database, and this function will return any of them, so this not usually what you want.

See deserialize() to recover a stored location.

station_code

a 4 letter METAR code

Returns

a weather station level location for the given station code, or NULL if none exists in the database

pub fn find_nearest_city(&self, lat: f64, lon: f64) -> Location

Finds the nearest city to the passed latitude and longitude, among the descendants of @self.

The given location must be at most a LocationLevel::Adm1 location. This restriction may be lifted in a future version.

Note that this function does not check if (@lat, @lon) fall inside @self, or are in the same region and time zone as the return value.

lat

Latitude, in degrees

lon

Longitude, in degrees

Returns

the city closest to (@lat, @lon), in the region or administrative district of @self. Finds the nearest city to the passed latitude and longitude, among the descendants of @self.

The given location must be at most a LocationLevel::Adm1 location. This restriction may be lifted in a future version.

Note that this function does not check if (@lat, @lon) fall inside @self, or are in the same region and time zone as the return value.

lat

Latitude, in degrees

lon

Longitude, in degrees

Returns

the city closest to (@lat, @lon), in the region or administrative district of @self.

pub fn find_nearest_city_full(
&self,
lat: f64,
lon: f64,
func: Option<Box_<dyn Fn(&Location) -> bool + 'static>>
) -> Location

Finds the nearest city to the passed latitude and longitude, among the descendants of @self.

Supports the use of own filter function to filter out locations.

Geocoding should be done on the application side if needed.

@self must be at most a LocationLevel::Adm1 location. This restriction may be lifted in a future version.

lat

Latitude, in degrees

lon

Longitude, in degrees

func

a function to iterate over the locations; the function must return TRUE to continue checking for the location, and FALSE to filter the location out

Returns

the city closest to (@lat, @lon), in the region or administrative district of @self with validation of filter function Finds the nearest city to the passed latitude and longitude, among the descendants of @self.

Supports the use of own filter function to filter out locations.

Geocoding should be done on the application side if needed.

@self must be at most a LocationLevel::Adm1 location. This restriction may be lifted in a future version.

lat

Latitude, in degrees

lon

Longitude, in degrees

func

a function to iterate over the locations; the function must return TRUE to continue checking for the location, and FALSE to filter the location out

Returns

the city closest to (@lat, @lon), in the region or administrative district of @self with validation of filter function

pub fn city_name(&self) -> Option<GString>

Retrieves the city name for the given location.

For a GWEATHER_LOCATION_CITY or GWEATHER_LOCATION_DETACHED location, this method is equivalent to name().

For a GWEATHER_LOCATION_WEATHER_STATION location, this is equivalent to calling name() on the location’s parent.

For other locations this method will return NULL.

Returns

the city name of the location Retrieves the city name for the given location.

For a GWEATHER_LOCATION_CITY or GWEATHER_LOCATION_DETACHED location, this method is equivalent to name().

For a GWEATHER_LOCATION_WEATHER_STATION location, this is equivalent to calling name() on the location’s parent.

For other locations this method will return NULL.

Returns

the city name of the location

pub fn code(&self) -> Option<GString>

Gets the METAR station code associated with a GWEATHER_LOCATION_WEATHER_STATION location.

Returns

the location’s METAR station code Gets the METAR station code associated with a GWEATHER_LOCATION_WEATHER_STATION location.

Returns

the location’s METAR station code

pub fn coords(&self) -> (f64, f64)

Gets @self’s coordinates.

You must call has_coords() before calling this function.

Returns

latitude

the return location for the latitude

longitude

the return location for the longitude Gets @self’s coordinates.

You must call has_coords() before calling this function.

Returns

latitude

the return location for the latitude

longitude

the return location for the longitude

pub fn country(&self) -> Option<GString>

Gets the ISO 3166 country code of the given location.

For GWEATHER_LOCATION_WORLD and GWEATHER_LOCATION_REGION, this function returns NULL.

Returns

the location’s country code Gets the ISO 3166 country code of the given location.

For GWEATHER_LOCATION_WORLD and GWEATHER_LOCATION_REGION, this function returns NULL.

Returns

the location’s country code

pub fn country_name(&self) -> Option<GString>

Retrieves the country name for the given location.

For a GWEATHER_LOCATION_COUNTRY location, this is equivalent to name().

For a GWEATHER_LOCATION_REGION and GWEATHER_LOCATION_WORLD location, this method will return NULL.

For other location levels, this method will find the parent GWEATHER_LOCATION_COUNTRY and return its name.

Returns

the location’s country name Retrieves the country name for the given location.

For a GWEATHER_LOCATION_COUNTRY location, this is equivalent to name().

For a GWEATHER_LOCATION_REGION and GWEATHER_LOCATION_WORLD location, this method will return NULL.

For other location levels, this method will find the parent GWEATHER_LOCATION_COUNTRY and return its name.

Returns

the location’s country name

pub fn distance(&self, loc2: &Location) -> f64

Determines the distance in kilometers between @self and @loc2.

loc2

a second #GWeatherLocation

Returns

the distance between @self and @loc2. Determines the distance in kilometers between @self and @loc2.

loc2

a second #GWeatherLocation

Returns

the distance between @self and @loc2.

pub fn english_name(&self) -> Option<GString>

Gets the location’s name.

Returns

the location’s name Gets the location’s name.

Returns

the location’s name

pub fn english_sort_name(&self) -> Option<GString>

Gets the location’s name, in a representation useful for comparisons.

The “sort name” is the location’s name after having g_utf8_normalize() (with G_NORMALIZE_ALL) and g_utf8_casefold() called on it. You can use this to sort locations, or to comparing user input against a location name.

Returns

the sort name of the location Gets the location’s name, in a representation useful for comparisons.

The “sort name” is the location’s name after having g_utf8_normalize() (with G_NORMALIZE_ALL) and g_utf8_casefold() called on it. You can use this to sort locations, or to comparing user input against a location name.

Returns

the sort name of the location

pub fn level(&self) -> LocationLevel

Gets @self’s level, from LocationLevel::World, to LocationLevel::WeatherStation.

Returns

@self’s level Gets @self’s level, from LocationLevel::World, to LocationLevel::WeatherStation.

Returns

@self’s level

pub fn name(&self) -> Option<GString>

Gets the location’s name, localized into the current language.

Returns

the location’s name Gets the location’s name, localized into the current language.

Returns

the location’s name

pub fn parent(&self) -> Option<Location>

Gets the location’s parent.

Returns

the location’s parent Gets the location’s parent.

Returns

the location’s parent

pub fn sort_name(&self) -> Option<GString>

Gets the location’s name, localized into the current language, in a representation useful for comparisons.

The “sort name” is the location’s name after having g_utf8_normalize() (with G_NORMALIZE_ALL) and g_utf8_casefold() called on it. You can use this to sort locations, or to comparing user input against a location name.

Returns

the sort name of the location Gets the location’s name, localized into the current language, in a representation useful for comparisons.

The “sort name” is the location’s name after having g_utf8_normalize() (with G_NORMALIZE_ALL) and g_utf8_casefold() called on it. You can use this to sort locations, or to comparing user input against a location name.

Returns

the sort name of the location

pub fn timezone(&self) -> Option<TimeZone>

Gets the timezone associated with @self, if known.

Returns

the location’s timezone Gets the timezone associated with @self, if known.

Returns

the location’s timezone

pub fn timezone_str(&self) -> Option<GString>

Gets the timezone associated with @self, if known, as a string.

Returns

the location’s timezone as a string Gets the timezone associated with @self, if known, as a string.

Returns

the location’s timezone as a string

pub fn timezones(&self) -> Vec<TimeZone>

Gets an array of all timezones associated with any location under @self.

Use gweather_location_free_timezones() to free this array.

Returns

the timezones associated with the location Gets an array of all timezones associated with any location under @self.

Use gweather_location_free_timezones() to free this array.

Returns

the timezones associated with the location

pub fn has_coords(&self) -> bool

Checks if @self has valid latitude and longitude.

Returns

true if @self has valid latitude and longitude. Checks if @self has valid latitude and longitude.

Returns

true if @self has valid latitude and longitude.

pub fn has_timezone(&self) -> bool

Checks whether the location has a timezone.

Returns

true if the location has a timezone; false otherwise Checks whether the location has a timezone.

Returns

true if the location has a timezone; false otherwise

pub fn next_child(&self, child: Option<&Location>) -> Option<Location>

Allows iterating all children of a location.

Pass NULL to get the first child, and any child to get the next one.

Note that the reference to @child is taken, meaning iterating all children is as simple as:

⚠️ The following code is in c ⚠️

  g_autoptr (GWeatherLocation) child = NULL;
  while ((child = gweather_location_next_child (location, child)))
    {
       // Do something with child
    }

child

the next child

Returns

The next child, if one exists Allows iterating all children of a location.

Pass NULL to get the first child, and any child to get the next one.

Note that the reference to @child is taken, meaning iterating all children is as simple as:

⚠️ The following code is in c ⚠️

  g_autoptr (GWeatherLocation) child = NULL;
  while ((child = gweather_location_next_child (location, child)))
    {
       // Do something with child
    }

child

the next child

Returns

The next child, if one exists

pub fn serialize(&self) -> Variant

Transforms a #GWeatherLocation into a #GVariant, in a way that calling gweather_location_deserialize() will hold an equivalent #GWeatherLocation. The resulting variant can then be stored into GSettings or on disk. This call is only valid for cities, weather stations and detached locations. The format of the resulting #GVariant is private to libgweather, and it is subject to change. You should use the “v” format in GSettings, to ensure maximum compatibility with future versions of the library.

Returns

the serialization of @location. Transforms a #GWeatherLocation into a #GVariant, in a way that calling gweather_location_deserialize() will hold an equivalent #GWeatherLocation. The resulting variant can then be stored into GSettings or on disk. This call is only valid for cities, weather stations and detached locations. The format of the resulting #GVariant is private to libgweather, and it is subject to change. You should use the “v” format in GSettings, to ensure maximum compatibility with future versions of the library.

Returns

the serialization of @location.

pub fn world() -> Option<Location>

Obtains the shared GWeatherLocation of type GWEATHER_LOCATION_WORLD, representing a hierarchy containing all of the locations from the location data.

Returns

a GWEATHER_LOCATION_WORLD location, or NULL if the locations data could not be found or could not be parsed. Obtains the shared GWeatherLocation of type GWEATHER_LOCATION_WORLD, representing a hierarchy containing all of the locations from the location data.

Returns

a GWEATHER_LOCATION_WORLD location, or NULL if the locations data could not be found or could not be parsed.

impl Location

pub fn detect_nearest_city<P: FnOnce(Result<Location, Error>) + 'static>(
&self,
lat: f64,
lon: f64,
cancellable: Option<&impl IsA<Cancellable>>,
callback: P
)

Initializes geocode reversing to find place for (@lat, @lon) coordinates.

Calls the callback function passed by user when the result is ready.

The given location must be at most a LocationLevel::Adm1 location; this restriction may be lifted in a future version.

lat

Latitude, in degrees

lon

Longitude, in degrees

cancellable

a cancellable instance

callback

callback function Initializes geocode reversing to find place for (@lat, @lon) coordinates.

Calls the callback function passed by user when the result is ready.

The given location must be at most a LocationLevel::Adm1 location; this restriction may be lifted in a future version.

lat

Latitude, in degrees

lon

Longitude, in degrees

cancellable

a cancellable instance

callback

callback function

pub fn detect_nearest_city_future(
&self,
lat: f64,
lon: f64
) -> Pin<Box_<dyn Future<Output = Result<Location, Error>> + 'static>>